+++ packages/utils/src/index.ts
@@ -1,2 +1,0 @@
-export { getDepsRegistrationList } from './deps-register.js';
-export { transformManifest } from './extension.js';