+++ tools/simple-cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @simple-data-open/simple-cli
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @simple-data-open/utils@0.0.20
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @simple-data-open/utils@0.0.19
+
 ## 0.0.18
 
 ### Patch Changes